Drill-specific changes: added a general class for Date/Time/Timestamp literals (TimestampString, DateString, TimeString) to avoid class cast exceptions.
In case of "date + interval" situation, for example:

where
  o.o_orderdate >= date '1996-10-01'
  and o.o_orderdate < date '1996-10-01' + interval '3' month

Drill will cast DATE to TIMESTAMP and we can face with ClassCastException like "ClassCastException: org.apache.calcite.util.DateString cannot be cast to org.apache.calcite.util.TimestampString" when Calcite will compare DATE ('1996-10-01') with TIMESTAMP ('1996-10-01' + interval '3'). To avoid this situation added a general class for Date/Time/Timestamp literals to make them comparable to each other.

Also added comparing NlsString with Date/Time/Timestamp strings to avoid ClassCastExceptoin in queries like:

WHERE birth_date BETWEEN '1920-01-01' AND cast('1931-01-01' AS DATE)

In this case, '1920-01-01' is NlsString and we got ClassCastExceptoin when we tried to compare it with Date '1931-01-01'.