[CALCITE-2291] [part-2] Fixes for support dynamic row type in UNNEST.
This commit has following additional changes:
a) Reverting the type for SqlUnnestOperator and Uncollect to ANY.
b) For the validation phase to pass use a place holder name for SqlUnnestOperator "$unnest"
c) Use the same name as the input name for Uncollect Node.
d) Fixing the converter tests to expect "$unnest".